Rename loading to loader. More sensible name.
