+++ src/index.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {audio as Audio} from './loading/audio';
-import {ballTrangle as BallTrangle } from './loading/ball-triangle';
-import {bars as Bars} from './loading/bars';
-import {circles as Circles} from './loading/circles';
-import {grid as Grid } from './loading/grid';
-import {hearts as Hearts} from './loading/hearts';
-import {oval as Oval } from './loading/oval';
-import {puff as Puff} from './loading/puff';
-import {rings as Rings} from './loading/rings';
-import {tailSpin as TailSpin} from './loading/tail-spin';
-import {threeDots as ThreeDots } from './loading/three-dots';
+import {audio as Audio} from './loader/audio';
+import {ballTrangle as BallTrangle } from './loader/ball-triangle';
+import {bars as Bars} from './loader/bars';
+import {circles as Circles} from './loader/circles';
+import {grid as Grid } from './loader/grid';
+import {hearts as Hearts} from './loader/hearts';
+import {oval as Oval } from './loader/oval';
+import {puff as Puff} from './loader/puff';
+import {rings as Rings} from './loader/rings';
+import {tailSpin as TailSpin} from './loader/tail-spin';
+import {threeDots as ThreeDots } from './loader/three-dots';
 export default class Loader extends React.Component {
   static propTypes = {
       color: PropTypes.string,
