+++ src/loader/bars.js
