+++ src/loader/hearts.js
