+++ src/loader/rings.js
